requirements.txt
requests
pandas
python-dotenv
supabase
fastapi
uvicorn
python-multipart
